depart x y fixed

+++ includes/cub3d.h
@@ -37,21 +37,21 @@
 typedef struct	s_ray {
 	int		rx; //width de la fenetre
 	int		ry; //height de la fenetre
-	float	posx; //position x du joueur
-	float	posy; //position y du joueur
-	float	dirx; //vecteur de direction (commence à -1 pour N, 1 pour S, 0 sinon)
-	float	diry; //vecteur de direction (commence à -1 pour W, 1 pour E, 0 sinon)
-	float	planx; //vecteur du plan (commence à 0.66 pour E, -0.66 pour W, 0 sinon)
-	float	plany; //vecteur du plan (commence à 0.66 pour N, -0.66 pour S, 0 sinon)
-	float	raydirx; //calcul de direction x du rayon
-	float	raydiry; //calcul de direction y du rayon
-	float	camerax; //point x sur la plan camera : Gauche ecran = -1, milieu = 0, droite = 1
+	double	posx; //position x du joueur
+	double	posy; //position y du joueur
+	double	dirx; //vecteur de direction (commence à -1 pour N, 1 pour S, 0 sinon)
+	double	diry; //vecteur de direction (commence à -1 pour W, 1 pour E, 0 sinon)
+	double	planx; //vecteur du plan (commence à 0.66 pour E, -0.66 pour W, 0 sinon)
+	double	plany; //vecteur du plan (commence à 0.66 pour N, -0.66 pour S, 0 sinon)
+	double	raydirx; //calcul de direction x du rayon
+	double	raydiry; //calcul de direction y du rayon
+	double	camerax; //point x sur la plan camera : Gauche ecran = -1, milieu = 0, droite = 1
 	int		mapx; // coordonée x du carré dans lequel est pos
 	int		mapy; // coordonnée y du carré dans lequel est pos
-	float	sidedistx; //distance que le rayon parcours jusqu'au premier point d'intersection vertical (=un coté x)
-	float	sidedisty; //distance que le rayon parcours jusqu'au premier point d'intersection horizontal (= un coté y)
-	float	deltadistx; //distance que rayon parcours entre chaque point d'intersection vertical
-	float	deltadisty; //distance que le rayon parcours entre chaque point d'intersection horizontal
+	double	sidedistx; //distance que le rayon parcours jusqu'au premier point d'intersection vertical (=un coté x)
+	double	sidedisty; //distance que le rayon parcours jusqu'au premier point d'intersection horizontal (= un coté y)
+	double	deltadistx; //distance que rayon parcours entre chaque point d'intersection vertical
+	double	deltadisty; //distance que le rayon parcours entre chaque point d'intersection horizontal
 	int		stepx; // -1 si doit sauter un carre dans direction x negative, 1 dans la direction x positive
 	int		stepy; // -1 si doit sauter un carre dans la direction y negative, 1 dans la direction y positive
 	int		hit; // 1 si un mur a ete touche, 0 sinon
@@ -84,8 +84,8 @@ typedef struct s_map
 	long	c;
 	int		height;
 	int		width;
-	float	pos_x;
-	float	pos_y;
+	double	pos_x;
+	double	pos_y;
 }			t_map;
 
 typedef struct s_player
